+++ cpp/open3d/visualization/rendering/Open3DScene.cpp
@@ -31,6 +31,7 @@
 #include "open3d/geometry/LineSet.h"
 #include "open3d/geometry/PointCloud.h"
 #include "open3d/geometry/TriangleMesh.h"
+#include "open3d/visualization/gui/Application.h"
 #include "open3d/visualization/rendering/Material.h"
 #include "open3d/visualization/rendering/Scene.h"
 #include "open3d/visualization/rendering/View.h"
@@ -119,6 +120,8 @@ Open3DScene::Open3DScene(Renderer& renderer) : renderer_(renderer) {
     view_ = scene->AddView(0, 0, 1, 1);
     scene->SetBackground({1.0f, 1.0f, 1.0f, 1.0f});
 
+    SetLighting(LightingProfile::MED_SHADOWS, {0.577f, -0.577f, -0.577f});
+
     RecreateAxis(scene, bounds_, false);
 }
 
@@ -154,6 +157,56 @@ void Open3DScene::SetBackground(const Eigen::Vector4f& color,
     scene->SetBackground(color, image);
 }
 
+void Open3DScene::SetLighting(LightingProfile profile,
+                              const Eigen::Vector3f& sun_dir) {
+    auto scene = renderer_.GetScene(scene_);
+
+    if (profile != LightingProfile::HARD_SHADOWS) {
+        if (scene->GetIndirectLight().empty()) {
+            auto path = gui::Application::GetInstance().GetResourcePath();
+            scene->SetIndirectLight(std::string(path) + "/default");
+        }
+    }
+
+    Eigen::Vector3f sun_color(1.0f, 1.0f, 1.0f);
+
+    // These intensities have been chosen so that a white object on a white
+    // background is clearly visible even when the highlight is next to the
+    // background. Increasing the intensities much more make the highlight's
+    // white too similar to the background's white.
+    switch (profile) {
+        case LightingProfile::HARD_SHADOWS:
+            scene->EnableIndirectLight(false);
+            scene->EnableDirectionalLight(true);
+            scene->SetDirectionalLight(sun_dir, sun_color, 100000);
+            break;
+        case LightingProfile::DARK_SHADOWS:
+            scene->EnableIndirectLight(true);
+            scene->EnableDirectionalLight(true);
+            scene->SetIndirectLightIntensity(5000);
+            scene->SetDirectionalLight(sun_dir, sun_color, 85000);
+            break;
+        default:
+        case LightingProfile::MED_SHADOWS:
+            scene->EnableIndirectLight(true);
+            scene->EnableDirectionalLight(true);
+            scene->SetIndirectLightIntensity(7500);
+            scene->SetDirectionalLight(sun_dir, sun_color, 70000);
+            break;
+        case LightingProfile::SOFT_SHADOWS:
+            scene->EnableIndirectLight(true);
+            scene->EnableDirectionalLight(true);
+            scene->SetIndirectLightIntensity(15000);
+            scene->SetDirectionalLight(sun_dir, sun_color, 35000);
+            break;
+        case LightingProfile::NO_SHADOWS:
+            scene->EnableIndirectLight(true);
+            scene->SetIndirectLightIntensity(20000);
+            scene->EnableDirectionalLight(false);
+            break;
+    }
+}
+
 void Open3DScene::ClearGeometry() {
     auto scene = renderer_.GetScene(scene_);
     for (auto& g : geometries_) {
@@ -172,7 +225,7 @@ void Open3DScene::ClearGeometry() {
 
 void Open3DScene::AddGeometry(
         const std::string& name,
-        std::shared_ptr<const geometry::Geometry3D> geom,
+        const geometry::Geometry3D* geom,
         const Material& mat,
         bool add_downsampled_copy_for_fast_rendering /*= true*/) {
     size_t downsample_threshold = SIZE_MAX;