+++ cpp/open3d/visualization/rendering/Open3DScene.h
@@ -65,6 +65,16 @@ public:
     void SetBackground(const Eigen::Vector4f& color,
                        std::shared_ptr<geometry::Image> image = nullptr);
 
+    enum class LightingProfile {
+        HARD_SHADOWS,
+        DARK_SHADOWS,
+        MED_SHADOWS,
+        SOFT_SHADOWS,
+        NO_SHADOWS
+    };
+
+    void SetLighting(LightingProfile profile, const Eigen::Vector3f& sun_dir);
+
     /// Sets the maximum number of points before AddGeometry also adds a
     /// downsampled point cloud with number of points, used when rendering
     /// speed is important.
@@ -76,7 +86,7 @@ public:
     void ClearGeometry();
     /// Adds a geometry with the specified name. Default visible is true.
     void AddGeometry(const std::string& name,
-                     std::shared_ptr<const geometry::Geometry3D> geom,
+                     const geometry::Geometry3D* geom,
                      const Material& mat,
                      bool add_downsampled_copy_for_fast_rendering = true);
     // Note: we can't use shared_ptr here, as we might be given something