+++ cpp/open3d/visualization/utility/Draw.h
@@ -24,31 +24,52 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#include "pybind/visualization/visualization.h"
+#pragma once
 
-#include "pybind/visualization/gui/gui.h"
-#include "pybind/visualization/rendering/rendering.h"
+#include <vector>
+
+#include "open3d/visualization/visualizer/O3DVisualizer.h"
 
 namespace open3d {
 namespace visualization {
 
-void pybind_visualization(py::module &m) {
-    py::module m_visualization = m.def_submodule("visualization");
-    pybind_renderoption(m_visualization);
-    pybind_viewcontrol(m_visualization);
-    pybind_visualizer(m_visualization);
-    pybind_visualization_utility(m_visualization);
-    pybind_renderoption_method(m_visualization);
-    pybind_viewcontrol_method(m_visualization);
-    pybind_visualizer_method(m_visualization);
-    pybind_visualization_utility_methods(m_visualization);
+struct DrawObject {
+    std::string name;
+    std::shared_ptr<geometry::Geometry3D> geometry;
+    std::shared_ptr<t::geometry::Geometry> tgeometry;
+    bool is_visible;
 
-#ifdef BUILD_GUI
-    rendering::pybind_rendering(m_visualization);
-    gui::pybind_gui(m_visualization);
-    pybind_o3dvisualizer(m_visualization);
-#endif
-}
+    DrawObject(const std::string &n,
+               std::shared_ptr<geometry::Geometry3D> g,
+               bool vis = true);
+    DrawObject(const std::string &n,
+               std::shared_ptr<t::geometry::Geometry> tg,
+               bool vis = true);
+};
+
+struct DrawAction {
+    std::string name;
+    std::function<void(visualizer::O3DVisualizer &)> callback;
+};
+
+void Draw(const std::vector<std::shared_ptr<geometry::Geometry3D>> &geometries,
+          const std::string &window_name = "Open3D",
+          int width = 1024,
+          int height = 768,
+          const std::vector<DrawAction> &actions = {});
+
+void Draw(
+        const std::vector<std::shared_ptr<t::geometry::Geometry>> &tgeometries,
+        const std::string &window_name = "Open3D",
+        int width = 1024,
+        int height = 768,
+        const std::vector<DrawAction> &actions = {});
+
+void Draw(const std::vector<DrawObject> &objects,
+          const std::string &window_name = "Open3D",
+          int width = 1024,
+          int height = 768,
+          const std::vector<DrawAction> &actions = {});
 
 }  // namespace visualization
 }  // namespace open3d